Code Review done to make it complaint with following two papers
1. Butz, Martin V., and Stewart W. Wilson. "An algorithmic description of XCS." International Workshop on Learning Classifier Systems. Springer, Berlin, Heidelberg, 2000.
2. Wilson, Stewart W. "Get real! XCS with continuous-valued inputs." International Workshop on Learning Classifier Systems. Springer, Berlin, Heidelberg, 1999.

Notes:
Code review done with respect to XCSR (real valued input) for single step scenario (non -dynamic).
Action set subsumption is not implemented for XCSR. Rest is as per the paper. Currently action set subsumption is disabled in parameters. If required we will implement it for real valued input.
GA and subsumption for ternary input not reviewed yet as we are not working on this type of problem yet. Can be reviewed when required.

- Completed code review of covering function
- Crossover and Mutation routines are implemented from scratch as per paper
There has been some small changes done to the code.
- Mutation probability parameter was not being used. it is corrected now
- Action mutation was not implemented. It is included now
- Formula for calculating error for a classifier is updated as per paper